pyproject.toml
[build-system]
requires = ["poetry", "poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "qermit"
version = "0.5.0"
description = "Python package for quantum error mitigation."
authors = [
    "Daniel Mills <daniel.mills@quantinuum.com>",
    "Silas Dilkes <silas.dilkes@quantinuum.com>",
    "Cristina Cirstoiu <cristina.cirstoiu@quantinuum.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4"
pytket-qiskit = "^0.50.0"
matplotlib = "^3.8.3"

pytest = { version = "^8.1.1", optional = true }
mypy = { version = "^1.9.0", optional = true }
sphinx = { version = "^8.1.3", optional = true }
pytest-cov = { version = "^6.0.0", optional = true }
qiskit-ibm-provider = { version = "^0.10.0", optional = true }
pytket-quantinuum = { version = "^0.41.0", optional = true }
ruff = { version = "^0.8.0", optional = true }
furo = { version = "^2024.8.6", optional = true }
myst-nb = { version = "^1.1.2", optional = true}
sphinx-autodoc-typehints = {version = "^2.5.0", optional = true}

[tool.poetry.extras]
tests = [
    "pytest",
    "mypy",
    "ruff",
    "pytest-cov",
    "qiskit-ibm-provider",
    "pytket-quantinuum",
]
docs = [
    "sphinx",
    "furo",
    "pytket-quantinuum",
    "myst-nb",
    "sphinx-autodoc-typehints",
]

[tool.ruff.lint]
select = ["I001"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
